+++ src/main/java/sysu/newchain/consensus/server/pbft/msg/PrepareMsg.java
@@ -47,11 +47,11 @@ public class PrepareMsg extends MsgBuilder<PreparePb.Builder>{
 	}
 	
 	public void setView(long view) {
-		getBuilder().setView(view);
+		getBuilder().setView((int) view);
 	}
 	
 	public void setSeqNum(long seqNum) {
-		getBuilder().setSeqNum(seqNum);
+		getBuilder().setSeqNum((int) seqNum);
 	}
 	
 	public void setDigestOfBlock(byte[] digest) {
